+^ PE - added long section names support & fixed several warnings

+++ Types/PE/src/PanelResources.cpp
@@ -76,7 +76,7 @@ void Panels::Resources::SaveCurrentResource()
     auto res = AppCUI::Dialogs::FileDialog::ShowSaveFileWindow(tmp, "", "");
     if (res.has_value())
     {
-        auto buf = this->win->GetObject()->GetData().CopyToBuffer(r->Start, r->Size);
+        auto buf = this->win->GetObject()->GetData().CopyToBuffer(r->Start, (uint32) r->Size);
         if (AppCUI::OS::File::WriteContent(res.value(), BufferView{ buf }) == false)
         {
             AppCUI::Dialogs::MessageBox::ShowError("Error", "Fail to create file !");
